src/main/java/com/desafio/desafiojava/controller/PessoaController.java
package com.desafio.desafiojava.controller;

import com.desafio.desafiojava.entity.Pessoa;
import com.desafio.desafiojava.service.PessoaService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.Optional;

@RestController
@RequestMapping("/api/pessoas")
public class PessoaController {

    private final PessoaService pessoaService;

    @Autowired
    public PessoaController(PessoaService pessoaService) {
        this.pessoaService = pessoaService;
    }

    // Route to create a new person
    @PostMapping
    public ResponseEntity<Pessoa> create(@RequestBody Pessoa pessoa) {
        Pessoa newPessoa = pessoaService.save(pessoa);
        return new ResponseEntity<>(newPessoa, HttpStatus.CREATED);
    }

    // Route to update an existing person
    @PutMapping("/{id}")
    public ResponseEntity<Pessoa> update(
            @PathVariable Long id, @RequestBody Pessoa updatedPessoa) {
        try {
            Pessoa pessoa = pessoaService.update(id, updatedPessoa);
            return ResponseEntity.ok(pessoa);
        } catch (RuntimeException e) {
            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
        }
    }

    // Route to retrieve all people
    @GetMapping
    public ResponseEntity<List<Pessoa>> getAll() {
        List<Pessoa> pessoas = pessoaService.findAll();
        return ResponseEntity.ok(pessoas);
    }

    // Route to retrieve a person by ID
    @GetMapping("/{id}")
    public ResponseEntity<Pessoa> getById(@PathVariable Long id) {
        Optional<Pessoa> pessoa = pessoaService.findById(id);
        return pessoa.map(ResponseEntity::ok)
                .orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND).body(null));
    }
}
